+++ CHANGELOG.md
@@ -9,9 +9,9 @@
 ### Fixes
 - Dynamic buffer recovery check fix
 
-### Core Updates
+### Internal Updates
 - Abstract server and I/O service
-- Use `shared_from_this` to keep reference count in the handler
+- Use shared pointer to keep reference for the session in the handler
 
 ## [2.2.1] - 19-03-2018
 ### Extensions API