+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change Log
 
-## [Unreleased]
+## [2.3.0] - 24-03-2018
 ### Extensions API
 - Added `reset` in `DispatchErrorExtension` which is called when previously failed is no longer failing